+++ integrations/clickup/CHANGELOG.md
@@ -11,4 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- Added Jira integration with support for projects and issues (0.1.0)
+- Added Clickup integration with support for projects and issues (0.1.0)